session7_var_autodecoder/aws_deploy/requirements.txt
numpy==1.18.5
onnxruntime==1.4.0
Pillow==7.2.0
requests_toolbelt
matplotlib==3.2.2
